Fixed MSVC warning in ScriptParser.
git-svn-id: http://www.bci2000.org/svn/trunk@4012 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/core/Operator/OperatorLib/ScriptParser/ScriptParser.cpp
@@ -1507,7 +1507,7 @@ yyreturn:
     streampos start = is.tellg();
     is >> ParserToken();
     streamsize length = is.tellg() - start;
-    pLval->token = new Token( *p, length );
+    pLval->token = new Token( *p, static_cast<int>( length ) );
     is.clear();
     is.seekg( start );
     is.read( pLval->token->Data(), length );
